Integrate QAConsole scripts into single class
